+++ Samples~/Scripts/GameManagerToolkit.cs
@@ -1,15 +1,21 @@
+using System.Collections;
 using UnityEngine;
 
 namespace MyUnityPackage.Toolkit
 {
     public class GameManagerToolkit : MonoBehaviour
     {
+     
         private void Start()
         {
             SetupGame();
             ServiceLocator.GetService<AudioPlaybackService>().PlayMusic((AudioClip)Resources.Load("Music"));
+            
+           
+          
         }
 
+      
         private void SetupGame()
         {
             AudioManager.Initialize();