+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.0.4] - 2025-11-06
+### Added
+- Add MUP_ObjectPool Component
+- Add MUP_ObjectPoolLocator Component
+
+### Removed
+- Animation System(move to another package)
+
 ## [1.0.3] - 2025-09-18
-### New Release
+### Refactor
 - Clear project & add docs
+- Changed Logger
+
+### Fixed
+- ServiceLocator
+- Sound Bug
+
+## [1.0.3] - 2025-07-11
+### Fixed
+- Fixed Samples
+
+[1.0.4]: https://github.com/AlexandreFlanquart/UnityPackage-Toolkit/releases/tag/1.0.4
+[1.0.3]: https://github.com/AlexandreFlanquart/UnityPackage-Toolkit/releases/tag/1.0.3
+[1.0.2]: https://github.com/AlexandreFlanquart/UnityPackage-Toolkit/releases/tag/1.0.2